refactor(core): move doctor line formatting and status markers below their callers
`formatDoctorCheckSummaryLine`, `formatDoctorCheckDetailLines`, and
`formatCliStatusMarker` are shared by the daemon doctor handlers, the CLI
replay-test reporter, and the doctor CLI output formatter. Owning them in
`src/daemon/handlers/` forced `src/commands` to reach up into `src/daemon`
through a dynamic import; `src/core` sits below every caller.
