+++ src/core/doctor-output.ts
